refactor: level module
