+++ src/level/components/level_tag.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Component;
+
+#[derive(Component)]
+pub struct LevelTag;
