+++ src/level/mod.rs
@@ -1,1 +1,3 @@
+pub mod components;
+pub mod load_level;
 pub mod plugin;
